'15.09.02Update
 점수 올라간ㄴ거 막음


git-svn-id: https://dev.naver.com/svn/eotktjdrkf@79 29f03ed2-db14-0510-ae99-898e00d8f4b7

+++ FrameWork_03/app/src/main/java/teamwarpcbstuido/IESEKI/org/SpriteAnimation.java
@@ -14,7 +14,7 @@ public class SpriteAnimation extends GraphicObject {
     private int mNoOfFrames;
     private int mCurrentFrame;
 
-    private long mFrameTimer;
+    private float mFrameTimer;
 
     private int mSpriteHeight;
     private int mSpriteWidth;
@@ -70,7 +70,7 @@ public class SpriteAnimation extends GraphicObject {
 
 
 
-    public void SpriteUpdate(long GameTime) {
+    public void SpriteUpdate(float GameTime) {
         if (!mbEnd) {
             if (GameTime > mFrameTimer + mFPS) {
                 mFrameTimer = GameTime;
